+++ tests/v2/templates.spec.js
@@ -12,10 +12,9 @@ test('Use template page', async ({ page }) => {
     await test.step('Check workflow not from templates', async () => {
         await page.getByRole('button', { name: 'Workflow properties' }).click();
         const modal = await waitModal(page);
-        // FIXME: the test below is not really working
         await expect(
             modal.getByText('This workflow comes from a template')
-        ).toHaveCount(0);
+        ).not.toBeVisible();
 
         await modal.getByRole(
             'textbox', { name: 'Workflow description' }
@@ -65,7 +64,7 @@ test('Use template page', async ({ page }) => {
     });
     
     await expect(page).toHaveURL(/\/v2\/templates\?template_id=/);
-    const url = new URL(page.url());
+    let url = new URL(page.url());
     const templateId = url.searchParams.get('template_id');
 
     await test.step('Check template info modal', async () => {
@@ -99,13 +98,15 @@ test('Use template page', async ({ page }) => {
         await waitModalClosed(page);
     });
 
+    const newDescription = 'New description.'
+
     await test.step('Edit template', async () => {
 
         await page.getByRole('button', { name: 'Edit' }).click();
         const modal = await waitModal(page);
 
         await modal.getByLabel('User Group').selectOption({ label: undefined });
-        await page.getByRole('textbox', { name: 'Description' }).fill('New description.');
+        await page.getByRole('textbox', { name: 'Description' }).fill(newDescription);
         await modal.getByRole('button', { name: 'Save' }).click();
         await waitModalClosed(page);
 
@@ -113,7 +114,7 @@ test('Use template page', async ({ page }) => {
         await waitModal(page);
         const values = page.locator('.list-group-item:not(.text-bg-light)');
         await expect(values.nth(4).locator('span')).toHaveText('-');
-        await expect(values.nth(5).locator('span')).toHaveText('New description.');
+        await expect(values.nth(5).locator('span')).toHaveText(newDescription);
 
         await modal.getByRole('button', { name: 'Close' }).click();
         await waitModalClosed(page);
@@ -122,9 +123,7 @@ test('Use template page', async ({ page }) => {
     let fileName;
 
     await test.step('Download and delete template', async () => {
-        const rows = page.locator('tbody tr');
-        await expect(await rows.count()).toBe(1)
-
+        await expect(page.getByRole('row')).toHaveCount(2);
         const downloadPromise = page.waitForEvent('download');
         await page.getByRole('button', { name: 'Download' }).click();
         const download = await downloadPromise;
@@ -137,7 +136,7 @@ test('Use template page', async ({ page }) => {
         await page.getByRole('button', { name: 'Confirm' }).click();
         await waitModalClosed(page);
         await expect(page).toHaveURL(`/v2/templates?template_id=${templateId}`);
-        await expect(await rows.count()).toBe(0)
+        await expect(page.getByRole('row')).toHaveCount(1);
 
         const applyButton = await page.getByRole('button', { name: 'Apply' });
         await expect(applyButton).toBeDisabled();
@@ -149,43 +148,120 @@ test('Use template page', async ({ page }) => {
         await expect(page).toHaveURL(`/v2/templates`);
     });
 
-    await test.step('Upload template', async () => {
+    const newTemplateName = Math.random().toString(36).substring(7);
+
+    await test.step('Upload templates', async () => {
 
         const applyButton = await page.getByRole('button', { name: 'Apply' });
         await expect(applyButton).toBeDisabled();
         const resetButton = await page.getByRole('button', { name: 'Reset' });
         await expect(resetButton).toBeDisabled();
         
-        const nameInput = page.locator('#searchTemplateName');
+        const nameInput = page.getByRole('textbox', { name: 'Name' });
         await expect(nameInput).toHaveValue('');
         await nameInput.fill(workflow.name);
+        
         await expect(applyButton).toBeEnabled();
         await applyButton.click()
         await waitPageLoading(page);
 
         await expect(page).toHaveURL(`/v2/templates?name=${workflow.name}`);
-        await expect(page.locator('tbody tr')).toHaveCount(0);
+        await expect(page.getByRole('row')).toHaveCount(1);
 
         await page.getByRole('button', { name: 'Import' }).click();
-        await page.locator('input#templateFile').setInputFiles(fileName);
+        await page.getByLabel('Select a file').setInputFiles(fileName);
         await page.getByRole('button', { name: 'Import template' }).click();
         await waitModalClosed(page);
         
-        await expect(page).toHaveURL(`/v2/templates?name=${workflow.name}`);
-        await expect(page.locator('tbody tr')).toHaveCount(1);
+        await expect(page).toHaveURL(/\/v2\/templates\?template_id=/);
+        await expect(page.getByRole('row')).toHaveCount(2);
+        // version is not a combobox now
+        await expect(
+            page.getByRole('row', {name: workflow.name})
+            .getByRole('combobox').locator('option')
+        ).toHaveCount(0);
 
         await page.getByRole('button', { name: 'Import' }).click();
-        await page.locator('input#templateFile').setInputFiles(fileName);
+        await page.getByLabel('Select a file').setInputFiles(fileName);
         await page.getByRole('button', { name: 'Import template' }).click();
         await expect(page.getByText(
             `The current user already own a workflow template with name='${workflow.name}' and version=1`
         )).toBeVisible();
-        await page.locator('#templateVersion').fill("42");
+        // override version
+        await page.getByRole('spinbutton', {name: 'Override template version'}).fill('42')
         await page.getByRole('button', { name: 'Import template' }).click();
-        
-        //  TODO: check we have two versions now
-        //  TODO: override template name
+        await resetButton.click()
+        await page.getByRole('textbox', {name: 'Name'}).fill(workflow.name)
+        await applyButton.click()
+        await expect(page.getByRole('row')).toHaveCount(2);
+        await expect(
+            page.getByRole('row', {name: workflow.name})
+            .getByRole('combobox').locator('option')
+        ).toHaveCount(2);
+        // override name
+        await page.getByRole('button', { name: 'Import' }).click();
+        await page.getByLabel('Select a file').setInputFiles(fileName);
+        await page.getByRole('textbox', {name: 'Override template name'}).fill(newTemplateName);
+        await page.getByRole('button', { name: 'Import template' }).click();
+        await expect(page).toHaveURL(/\/v2\/templates\?template_id=/);
         fs.rmSync(fileName);
     });
 
+
+    await test.step('Create a workflow from a template', async () => {
+        await page.goto(`/v2/projects/${project.id}`);
+        await page.getByRole('button', { name: 'Create new workflow' }).click();
+        const modal = await waitModal(page);
+        await modal.getByLabel('Create from template').check();
+
+        const rows = modal.locator('tbody tr');
+        await expect(rows.first()).toBeVisible();
+
+        const count1 = await rows.count();
+        expect(count1).toBeGreaterThan(1);
+
+        const applyButton = await modal.getByRole('button', { name: 'Apply' });
+        await expect(applyButton).toBeDisabled();
+
+        
+        await modal.getByRole('textbox', { name: 'Name' }).nth(1).fill(newTemplateName);
+        await expect(applyButton).toBeEnabled();
+        await applyButton.click()
+        
+        await expect(modal.getByRole('row')).toHaveCount(2);
+
+        await modal.getByRole('button', { name: 'Select' }).click();
+        await expect(modal.getByText(
+            `Workflow with name='${workflow.name}' and project_id=${project.id} already exists.`
+        )).toBeVisible();
+        const workflowNameInput = modal.getByRole('textbox', { name: 'Workflow name' });
+        await workflowNameInput.fill(Math.random().toString(36).substring(7));
+        await workflowNameInput.blur();
+        await modal.getByRole('button', { name: 'Select' }).click();
+
+        await page.getByRole('button', { name: 'Workflow properties' }).click();
+        const modal2 = await waitModal(page);
+        await expect(
+            modal2.getByText('This workflow comes from a template')
+        ).toBeVisible();
+        await modal2.getByRole('button', { name: 'Close' }).click();
+        await waitModalClosed(page);
+
+        await page.getByRole('button', { name: 'Create template' }).click();
+        const modal3 = await waitModal(page);
+        // default: originalTemplate.name
+        await expect(
+            modal3.getByText('Template name')
+        ).toHaveValue(newTemplateName);
+        // default: 2
+        await expect(
+            modal3.getByText('Template version')
+        ).toHaveValue('2');
+        // default originalTemplate.description
+        await expect(
+            modal3.getByText('Template description')
+        ).toHaveValue(newDescription);
+        // default: null
+        await expect(modal3.getByLabel('User Group')).toHaveValue('');
+    });
 });